+++ src/modules-arch/module-scheme.js
@@ -9,9 +9,6 @@ export class ModuleScheme {
    * Build the scheme
    *
    * @param {Object} scheme
-   * @param {String} scheme.sidebarLabel
-   * Label that should be displayed in the sidebar. To make easier
-   * distinguishing of the current scheme.
    *
    * @param {String} scheme.appLogoUrl
    * URL of the image that should be displayed as an application logo
@@ -25,7 +22,6 @@ export class ModuleScheme {
     this._validateRawScheme(scheme)
 
     this._rawScheme = scheme
-    this._sidebarLabel = scheme.sidebarLabel
     this._appLogoUrl = scheme.appLogoUrl
     this._pages = scheme.pages
     this._importEnLocaleFile = scheme.importEnLocaleFile
@@ -35,7 +31,6 @@ export class ModuleScheme {
     this._validateCache()
   }
 
-  get sidebarLabel () { return this._sidebarLabel }
   get appLogoUrl () { return this._appLogoUrl }
   get pages () { return this._pages }
   get cache () { return this._cache }