+++ CHANGELOG.md
@@ -54,6 +54,8 @@ for further information about branching and tagging conventions.
 - Removed wallet providing to each module
 - Now displaying disabled investing reason using "MessageBox" component
 - Moved exceeding sale cap message to vuelidate error messages
+- Module schemes:
+  -  Removed schemeLabel
 
 #### Removed
 - Unused methods from `DocumentContainer` class